+++ src/dipdup/indexes/evm_subsquid.py
@@ -1,6 +1,5 @@
 import asyncio
 import random
-import time
 from abc import ABC
 from abc import abstractmethod
 from collections import defaultdict
@@ -16,6 +15,7 @@ from dipdup.config import SubsquidIndexConfigU
 from dipdup.config.evm import EvmContractConfig
 from dipdup.config.evm_node import EvmNodeDatasourceConfig
 from dipdup.context import DipDupContext
+from dipdup.datasources import IndexDatasource
 from dipdup.datasources.evm_node import NODE_LAST_MILE
 from dipdup.datasources.evm_node import EvmNodeDatasource
 from dipdup.datasources.evm_subsquid import SubsquidDatasource
@@ -25,7 +25,6 @@ from dipdup.index import Index
 from dipdup.index import IndexQueueItemT
 from dipdup.models.evm_subsquid import SubsquidMessageType
 from dipdup.package import DipDupPackage
-from dipdup.performance import metrics
 from dipdup.prometheus import Metrics
 
 IndexConfigT = TypeVar('IndexConfigT', bound=SubsquidIndexConfigU)
@@ -51,10 +50,10 @@ def get_sighash(package: DipDupPackage, method: str, to: EvmContractConfig | Non
 
 
 class SubsquidIndex(
+    Generic[IndexConfigT, IndexQueueItemT, DatasourceT],
     Index[IndexConfigT, IndexQueueItemT, DatasourceT],
     ABC,
-    Generic[IndexConfigT, IndexQueueItemT, DatasourceT],
-    message_type=SubsquidMessageType,
+    message_type=SubsquidMessageType,  # type: ignore[arg-type]
 ):
     def __init__(
         self,
@@ -79,23 +78,13 @@ class SubsquidIndex(
     @abstractmethod
     async def _synchronize_node(self, sync_level: int) -> None: ...
 
-    @abstractmethod
-    def _match_level_data(
-        self,
-        handlers: Any,
-        level_data: Any,
-    ) -> deque[Any]: ...
-
-    @abstractmethod
-    async def _call_matched_handler(
-        self,
-        handler_config: Any,
-        level_data: Any,
-    ) -> None: ...
-
     @property
     def node_datasources(self) -> tuple[EvmNodeDatasource, ...]:
         return self._node_datasources
+
+    @property
+    def datasources(self) -> tuple[IndexDatasource[Any], ...]:
+        return (self.datasource, *self.node_datasources)
 
     def get_random_node(self) -> EvmNodeDatasource:
         if not self._node_datasources:
@@ -199,37 +188,3 @@ class SubsquidIndex(
             await self._synchronize_subsquid(sync_level)
 
         await self._exit_sync_state(sync_level)
-
-    async def _process_level_data(
-        self,
-        level_data: Any,
-        sync_level: int,
-    ) -> None:
-        if not level_data:
-            return
-
-        batch_level = level_data[0].level
-        index_level = self.state.level
-        if batch_level <= index_level:
-            raise FrameworkException(f'Batch level is lower than index level: {batch_level} <= {index_level}')
-
-        self._logger.debug('Processing data of level %s', batch_level)
-        started_at = time.time()
-        matched_handlers = self._match_level_data(self._config.handlers, level_data)
-
-        total_matched = len(matched_handlers)
-        Metrics.set_index_handlers_matched(total_matched)
-        metrics[f'{self.name}:handlers_matched'] += total_matched
-        metrics[f'{self.name}:time_in_matcher'] += (time.time() - started_at) / 60
-
-        # NOTE: We still need to bump index level but don't care if it will be done in existing transaction
-        if not matched_handlers:
-            await self._update_state(level=batch_level)
-            return
-
-        started_at = time.time()
-        async with self._ctx.transactions.in_transaction(batch_level, sync_level, self.name):
-            for handler_config, data in matched_handlers:
-                await self._call_matched_handler(handler_config, data)
-            await self._update_state(level=batch_level)
-        metrics[f'{self.name}:time_in_callbacks'] += (time.time() - started_at) / 60