+++ src/dipdup/models/tezos_tzkt.py
@@ -720,9 +720,3 @@ class TzktEvent(Generic[EventType]):
 class TzktUnknownEvent:
     data: TzktEventData
     payload: Any | None
-
-
-@dataclass(frozen=True)
-class TzktRollbackMessage:
-    from_level: int
-    to_level: int