+++ src/dipdup/install.py
@@ -44,7 +44,7 @@ Documentation:         https://dipdup.io/docs
 GitHub:                https://github.com/dipdup-io/dipdup
 Discord:               https://discord.gg/aG8XKuwsQd
 """
-DIPDUP_LTS_VERSION = '6.5.15'
+DIPDUP_LTS_VERSION = '6.5.16'
 
 
 class Colors: